fix youtube engine, support webm by default

+++ GmediaFinder/functions.py
@@ -37,28 +37,19 @@ def get_url_data(url):
         return data
         
 def download_photo(img_url):
-        filename = os.path.basename(img_url)
-        if sys.platform == "win32":
-            file_path = os.path.join(tempfile.gettempdir(), filename)
-        else:
-            file_path = "%s%s" % ('/tmp/', filename)
-        if os.path.exists(file_path):
-            os.remove(file_path)
-        p = urllib.urlretrieve(img_url, file_path)
+    filename = os.path.basename(img_url)
+    if sys.platform == "win32":
+        file_path = os.path.join(tempfile.gettempdir(), filename)
+    else:
+        file_path = "/tmp/%s" % filename
+    if os.path.exists(file_path) and not os.path.isdir(file_path):
+        os.remove(file_path)
+    p = urllib.urlretrieve(img_url, file_path)
+    try:
         vid_pic = gtk.gdk.pixbuf_new_from_file_at_scale(p[0],100,100,1)
-	return vid_pic
-
-def get_codec(num):
-    codec=None
-    if re.match('5|34|35',num):
-        codec = "flv"
-    elif re.match('18|22|37|38',num):
-        codec= "mp4"
-    elif re.match('43|45',num):
-        codec= "webm"
-    elif re.match('17',num):
-        codec= "3gp"
-    return codec
+        return vid_pic
+    except:
+        return None
 
 class Downloader(threading.Thread):
     def __init__(self,gui,url, name, pbar, btnf, btn,btn_conv,btnstop,convert,label=''):
